requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch==2.3.0
transformers==4.40.0
Pillow==10.3.0